+++ changelog.md
@@ -4,6 +4,7 @@
 
 - structurizr-dsl: Fixes https://github.com/structurizr/java/issues/374 (!identifiers hierarchical isn't propagated when extending a workspace).
 - structurizr-dsl: Adds the ability to use the `group` keyword inside a component definition, to set the group name of that component.
+- structurizr-dsl: Adds the ability to use the `group` keyword inside the component finder strategy `forEach` block.
 
 ## 3.2.1 (10th December 2024)